+++ src/test/java/com/stephan/adventofcode/y2021/day1/Day1Test.java
@@ -18,7 +18,7 @@ public class Day1Test {
   @Test
   void canReturnPart1Result() {
     // When
-    int result = underTest.part1Result();
+    long result = underTest.part1Result();
 
     // Then
     assertThat(result).isEqualTo(7);
@@ -27,7 +27,7 @@ public class Day1Test {
   @Test
   void canReturnPart2Result() {
     // When
-    int result = underTest.part2Result();
+    long result = underTest.part2Result();
 
     // Then
     assertThat(result).isEqualTo(5);
